health: move cilium health logic from cilium-health/launch to pkg/health
This commit moves the cilium health logic from `cilium-health/launch` to
`pkg/health`. Nowadays, this logic is only used by the agent and not used
by the `cilium-health` binary.
